requirements.txt
Django>=4.2,<4.3
python-docx>=0.8.11
docx2python>=2.0.0